+++ test/unit_tests/test_yaml.py
@@ -4,6 +4,7 @@ import os
 
 import yaml
 from amazonia.classes.asg_config import InvalidAsgConfigError
+from amazonia.classes.lambda_config import InvalidLambdaConfigError
 from amazonia.classes.util import InsecureVariableError
 from amazonia.classes.yaml import Yaml, InvalidYamlValueError, InvalidYamlStructureError
 from amazonia.classes.yaml_fields import YamlFields
@@ -193,6 +194,19 @@ def test_invalid_min_max_asg():
     assert_raises(InvalidAsgConfigError, Yaml, **{'user_stack_data': invalid_min_max_stack_data,
                                                   'default_data': default_data})
 
+
+@with_setup(setup_resources)
+def test_invalid_lambda_memory_size():
+    """
+    Test the detection of a larger minimum that the provided maximum for an auto scaling unit
+    """
+    global default_data
+    invalid_lambda_memory_stack_data = open_yaml_file('test_yaml_invalid_lambda_memory.yaml')
+
+    assert_raises(InvalidLambdaConfigError, Yaml, **{'user_stack_data': invalid_lambda_memory_stack_data,
+                                                     'default_data': default_data})
+
+
 @with_setup(setup_resources)
 def test_bad_defaults():
     """
@@ -201,4 +215,4 @@ def test_bad_defaults():
     bad_default_data = open_yaml_file('test_yaml_bad_defaults.yaml')
     valid_stack_data = open_yaml_file('test_yaml_complete_valid.yaml')
     assert_raises(InvalidYamlStructureError, Yaml, **{'user_stack_data': valid_stack_data,
-                                                  'default_data': bad_default_data})
+                                                      'default_data': bad_default_data})